pyproject.toml
[project]
name = "Chunk"
version = "0.1.0"
dependencies = [
    "bs4>=0.0.2",
    "document>=1.0",
    "docx>=0.2.4",
    "groq>=0.18.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pydub>=0.25.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "speechrecognition>=3.14.1",
    "streamlit>=1.42.1",
]
